secure delete photos/id route

+++ routes/photosRoutes.js
@@ -93,6 +93,17 @@ router.patch('/photos/:photo_id', async (req, res) => {
 
 router.delete('/photos/:photo_id', async (req, res) => {
   try {
+    const decodedToken = jwt.verify(req.cookies.jwt, jwtSecret)
+    if (!decodedToken) {
+      throw new Error('Invalid authentication token')
+    }
+    const { photosRows } = await db.query(`
+      SELECT * FROM photos WHERE id = ${req.params.photo_id}
+      LEFT JOIN houses ON houses.house_id = photos.house_id
+    `)
+    if (photosRows.rows[0].user_id !== decodedToken.user_id) {
+      throw new Error('You are not authorized to delete photos for this house')
+    }
     let { rows } = await db.query(
       `DELETE FROM houses_photos WHERE id = ${req.params.photo_id} RETURNING *`
     )
